+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.52.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.52.0)
+_April 08, 2024_
+
 ## StreamChat
 ### 🐞 Fixed
 - Fix pinning messages with attachments stuck in sending state [#3116](https://github.com/GetStream/stream-chat-swift/pull/3116)